Nicolas Arias Code Challenge
Since the content of the results is not served directly in the HTML,
regular scraping techniques don't work. For this solution, I used a
headless browser library called puppeteer. It allows rendering web pages
and executing javascript so that the final DOM can be examined
regularly.

From there, it was just a matter of identifying the HTML
structure of the card results, relying mostly on the tag names and tag
hierarchy as well as a few element ids that weren't obfuscated or
randomly generated. The contents of the cards were identified and
parsed, and then converted to the required format.

I initially worked with the provided VanGogh paintings file, but when I
obtained the HTML for the other search results (Pixar Movies and Lord of
the Rings Cast), I realized that the structure had changed at some point
between the challenge was first created and when I solved it. This
caused the script to fail with fresh results. I resorted to getting a
fresh result for the Van Gogh paintings search, included in the
`van-gogh-paintings-2.html` file, and was able to get the script working
correctly with the three different searches.

The script was written in NodeJS, using Puppeteer as a dependency to
render the website correctly by running the embedded Javascript.

The script requires NodeJS and NPM to work, it can be executed by
running the following commands from the project's root directory:

```
$ npm install
$ node index.js
```

This will parse the van gogh paintings file and output the results to
the console. Alternatively, all of the resulting files can be found in
the `output` directory.

The tests can be executed by running the following commands from the
project's root directory:

```
$ npm install
$ npm test
```