google carousel image parser
Contains a parser class that looks for an HTML element with tag
`g-scrolling-carousel` (the carousel) and then its child anchor and
`g-img` tags.

The page has some JavaScript that needs to be run to properly fill
image sources. For this, Selenium is used with the expectation that
the user will have Selenium and Chrome installed on their machine.

PhantomJS was attempted but there were some encoding issues for
special characters like `Café`.

NodeJS and other solutions were considered but this was easier, in
the given time - plus, fewer lines to review this way.

Setup and test details in the new `README.md` file.